Avoid deleting dummy repository files unless needed
This practice not only breaks caching, but also causes issues with concurrent invocations of Gradle tasks (i.e. IDE refresh).

+++ common/src/main/java/net/neoforged/gradle/common/extensions/repository/IvyDummyRepositoryExtension.java
@@ -168,12 +168,16 @@ public abstract class IvyDummyRepositoryExtension implements ConfigurableDSLElem
 
         writeIvyMetadataFile(entry, jarFile, baseDir, metaFile);
 
-        Files.deleteIfExists(jarFile);
-        Files.createFile(jarFile);
+        if (!Files.isRegularFile(jarFile)) {
+            Files.deleteIfExists(jarFile);
+            Files.createFile(jarFile);
+        }
 
         final Path sourcesFile = entry.asSources().buildArtifactPath(getRepositoryDirectory().get().getAsFile().toPath());
-        Files.deleteIfExists(sourcesFile);
-        Files.createFile(sourcesFile);
+        if (!Files.isRegularFile(sourcesFile)) {
+            Files.deleteIfExists(sourcesFile);
+            Files.createFile(sourcesFile);
+        }
 
         writeDummyDependencyDataIfNeeded(entry);
     }
